Refactored FareType to Fare, removed original Fare class
