+++ src/main/java/com/ise/taxiapp/nav/Region.java
@@ -1,7 +1,7 @@
 package com.ise.taxiapp.nav;
 
 import com.ise.taxiapp.entities.Taxi;
-import com.ise.taxiapp.entities.fare.Fare;
+import com.ise.taxiapp.entities.Fare;
 
 public class Region {
     public Taxi findNearest(Fare fare) {
